Fixes, Tests
